Add tox and travis environments for variations with and without lxml.
Since we use either the integrated XML parsing of Python or the one
provided by the 'lxml' library, it is best to run our tests in both
setups. Especially since lxml supports additional functionality, so we
have to make sure not to rely on this in the code.

OWSLib has a similar testing environment.