Fix parsing of invalid XML documents.
If the user has lxml installed, we try parsing the XML documents with the
'recover' option to parse even invalid documents.

If lxml is not installed we have to rely on the default Python XML
library which does not support the 'recover' option. In this case we
cannot parse invalid XML documents and have to set the dataframe columns
relying on XML information to 'nan'. We warn the user that the returned
dataframe is incomplete.