Align everything to a values dict.
We cannot make the distinction between {code: value} (vocubulary)
and {value:definition} (values) reliably in XSD.

This means leaving the replacing of codes by values up to the user.
These are probably better added as a new column anyway, since datatypes
of codes and values often differ. Should be explained in the docs.